Move GetSyntaxTree to document snapshot
