+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/GoToDefinition/AbstractRazorComponentDefinitionService.cs
@@ -69,11 +69,8 @@ internal abstract class AbstractRazorComponentDefinitionService(
         {
             _logger.LogInformation($"Attempting to get definition from an attribute directly.");
 
-            var originCodeDocument = await documentSnapshot.GetGeneratedOutputAsync().ConfigureAwait(false);
-            var syntaxTree = await GetCSharpSyntaxTreeAsync(documentSnapshot, cancellationToken).ConfigureAwait(false);
-
             var range = await RazorComponentDefinitionHelpers
-                .TryGetPropertyRangeAsync(originCodeDocument, syntaxTree, attributeDescriptor.GetPropertyName(), _documentMappingService, _logger, cancellationToken)
+                .TryGetPropertyRangeAsync(documentSnapshot, attributeDescriptor.GetPropertyName(), _documentMappingService, _logger, cancellationToken)
                 .ConfigureAwait(false);
 
             if (range is not null)
@@ -88,6 +85,4 @@ internal abstract class AbstractRazorComponentDefinitionService(
         // at least then press F7 to go there.
         return VsLspFactory.DefaultRange;
     }
-
-    protected abstract ValueTask<SyntaxTree> GetCSharpSyntaxTreeAsync(IDocumentSnapshot documentSnapshot, CancellationToken cancellationToken);
 }
