+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/GoToDefinition/RazorComponentDefinitionHelpers.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Razor.Language.Syntax;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Razor.DocumentMapping;
 using Microsoft.CodeAnalysis.Razor.Logging;
+using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 using LspRange = Microsoft.VisualStudio.LanguageServer.Protocol.Range;
@@ -129,8 +130,7 @@ internal static class RazorComponentDefinitionHelpers
     }
 
     public static async Task<LspRange?> TryGetPropertyRangeAsync(
-        RazorCodeDocument codeDocument,
-        SyntaxTree csharpSyntaxTree,
+        IDocumentSnapshot documentSnapshot,
         string propertyName,
         IDocumentMappingService documentMappingService,
         ILogger logger,
@@ -148,7 +148,9 @@ internal static class RazorComponentDefinitionHelpers
         //      will error, but allowing them to Go To Def on that property regardless, actually helps
         //      them fix the error.
 
+        var csharpSyntaxTree = await documentSnapshot.GetCSharpSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
         var root = await csharpSyntaxTree.GetRootAsync(cancellationToken).ConfigureAwait(false);
+        var codeDocument = await documentSnapshot.GetGeneratedOutputAsync().ConfigureAwait(false);
 
         // Since we know how the compiler generates the C# source we can be a little specific here, and avoid
         // long tree walks. If the compiler ever changes how they generate their code, the tests for this will break
