+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/ProjectSystem/IDocumentSnapshot.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System.Diagnostics.CodeAnalysis;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.CodeAnalysis.Text;
@@ -19,6 +20,12 @@ internal interface IDocumentSnapshot
     Task<SourceText> GetTextAsync();
     Task<VersionStamp> GetTextVersionAsync();
     Task<RazorCodeDocument> GetGeneratedOutputAsync();
+
+    /// <summary>
+    /// Gets the Roslyn syntax tree for the generated C# for this Razor document
+    /// </summary>
+    /// <remarks>Using this from the LSP server side of things is not ideal. Use sparingly :)</remarks>
+    Task<SyntaxTree> GetCSharpSyntaxTreeAsync(CancellationToken cancellationToken);
 
     bool TryGetText([NotNullWhen(true)] out SourceText? result);
     bool TryGetTextVersion(out VersionStamp result);
